More seed data
Adding some more race result data for the rest of the
dirvers i have in the database. Adding a new track also
so that I can eventually add more data. Setting up a new
calc that will calculate the number of race day fastest laps
a driver has done in a season