Chart changes and clean up
cleaned up some of the code, but I want to come back
and absrasct some code into sepearte functions.

made the graphdata more automated from the data
i have brought back where I can

added the team mate of the driver you seletected to the
chart so that you can compare to the other guy in the same
car.